requirements.txt
streamlit~=1.16.0
pillow~=9.4.0
pandas~=1.5.2